requirements.txt
fastapi
pydantic
uvicorn
email-validator
python-jose
python-multipart
bcrypt
passlib
psycopg2-binary
SQLAlchemy
python-dotenv
pydantic-settings
pyotp
